requirements.txt
fastapi
uvicorn
mysql-connector-python
sqlalchemy
cryptography